Funcion daño del player y particulas de movimiento

+++ Assets/Scripts/Characters/Characters_Scriptable.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Characters", menuName = "Game/Characters", order = 1)]
 public class Characters_Scriptable : ScriptableObject
 {
-    public float maxLife;
-    public float currentLife;
+    public int maxLife;
+    public int currentLife;
     public float characSpeed;
